requirements.txt
fastapi
uvicorn
sqlalchemy
celery
redis
openai
python-dotenv
alembic